requirements.txt
psycopg2>=2.7,<3.0
ipython==7.7.0
numpy==1.17.0
biopython==1.74
celery==4.3.0
djangorestframework==3.10.2
django-cors-headers==3.1.0
Django==2.2.4
sqlparse==0.3.0
SQLAlchemy==1.3.7
gunicorn==19.9.0
psycopg2-binary==2.8.3
uWSGI==2.0.18